Switch to standard uint64_t & int64_t data types.

+++ src/qt/optionsmodel.cpp
@@ -166,11 +166,11 @@ QVariant OptionsModel::data(const QModelIndex & index, int role) const
         case ProxySocksVersion:
             return settings.value("nSocksVersion", 5);
         case SplitThreshold:
-            return settings.value("nSplitThreshold", DEF_SPLIT_AMOUNT/1000000);
+            return settings.value("nSplitThreshold", static_cast<qlonglong>(DEF_SPLIT_AMOUNT/1000000));
         case MaxSplitThreshold:
-            return settings.value("nMaxSplitThreshold", MAX_SPLIT_AMOUNT/1000000);
+            return settings.value("nMaxSplitThreshold", static_cast<qlonglong>(MAX_SPLIT_AMOUNT/1000000));
         case Fee:
-            return QVariant(nTransactionFee);
+            return QVariant(static_cast<qlonglong>(nTransactionFee));
         case DisplayUnit:
             return QVariant(nDisplayUnit);
         case DisplayAddresses:
@@ -249,7 +249,7 @@ bool OptionsModel::setData(const QModelIndex & index, const QVariant & value, in
         break;
         case Fee:
             nTransactionFee = value.toLongLong();
-            settings.setValue("nTransactionFee", nTransactionFee);
+            settings.setValue("nTransactionFee", static_cast<qlonglong>(nTransactionFee));
             emit transactionFeeChanged(nTransactionFee);
             break;
         case SplitThreshold:
